Extract time series interval helper
